fix and scope styling for monitoring view and endpoint list
