+++ src/ServicePulse.Host/vue/src/assets/main.css
@@ -1,7 +1,6 @@
 @import "bootstrap/dist/css/bootstrap.css";
 @import "bootstrap-icons/font/bootstrap-icons.css";
 @import "font-awesome/css/font-awesome.css";
-@import "navbar.css";
 @import "footer.css";
 @import "tabs.css";
 @import "dropdown.css";
