+++ src/ServicePulse.Host/vue/src/components/PageHeader.vue
@@ -105,6 +105,13 @@ const displayDanger = computed(() => {
 </template>
 
 <style scoped>
+@import "@/assets/navbar.css";
+
+.nav {
+  --bs-link-color: #9d9d9d;
+  --bs-link-hover-color: #fff;
+}
+
 .navbar > .container-fluid > div {
   margin: 0 1em;
 }
