+++ src/ServicePulse.Host/vue/src/components/configuration/HealthCheckNotifications.vue
@@ -185,6 +185,8 @@ onMounted(() => {
 </template>
 
 <style>
+@import "../list.css";
+
 .screen-intro {
   margin: 30px 0;
 }
@@ -279,18 +281,6 @@ onMounted(() => {
   list-style-type: none;
   margin: 0;
   padding: 0;
-}
-
-.lead {
-  -ms-word-wrap: break-word;
-  word-wrap: break-word;
-  color: #181919 !important;
-  font-size: 14px !important;
-  font-weight: bold;
-  margin-bottom: 3px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
 }
 
 p.endpoint-metadata {
