+++ src/ServicePulse.Host/vue/src/components/failedmessages/FailedMessageGroups.vue
@@ -146,14 +146,5 @@ onMounted(async () => {
 </template>
 
 <style>
-.lead {
-  word-wrap: break-word;
-  color: #181919 !important;
-  font-size: 1em !important;
-  font-weight: bold !important;
-  margin-bottom: 0.2em;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
+@import "../list.css";
 </style>
