+++ src/ServicePulse.Host/vue/src/components/failedmessages/MessageView.vue
@@ -482,6 +482,8 @@ onUnmounted(() => {
 </template>
 
 <style>
+@import "../list.css";
+
 h1.message-type-title {
   margin: 0 0 8px;
   font-size: 24px;
@@ -503,16 +505,5 @@ button img {
 
 .msg-tabs {
   margin-bottom: 20px;
-}
-
-.lead {
-  word-wrap: break-word;
-  color: #181919 !important;
-  font-size: 1em !important;
-  font-weight: bold !important;
-  margin-bottom: 0.2em;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
 }
 </style>
