+++ src/ServicePulse.Host/vue/src/components/list.css
@@ -1,0 +1,14 @@
+.lead {
+  word-wrap: break-word;
+  color: #181919;
+  font-size: 1em;
+  font-weight: bold;
+  margin-bottom: 0.2em;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+a {
+  cursor: pointer;
+}
