+++ src/ServicePulse.Host/vue/src/components/monitoring/MonitoringFilter.vue
@@ -1,0 +1,72 @@
+<script setup>
+import { ref, watch } from "vue";
+import GroupBy from "./MonitoringGroupBy.vue";
+import PeriodSelector from "./MonitoringHistoryPeriod.vue";
+import { useMonitoringStore } from "@/stores/MonitoringStore";
+import { useRoute } from "vue-router";
+
+const monitoringStore = useMonitoringStore();
+const filterString = ref(monitoringStore.filterString);
+const route = useRoute();
+
+watch(route, () => monitoringStore.setHistoryPeriod(route.params.historyPeriod), { deep: true, immediate: true, flush: "pre" });
+watch(filterString, async (newValue) => {
+  await monitoringStore.updateFilterString(newValue);
+  filterString.value = monitoringStore.filterString;
+});
+</script>
+
+<template>
+  <div class="filter-group filter-monitoring">
+    <PeriodSelector />
+    <GroupBy />
+    <div class="filter-input">
+      <input type="text" placeholder="Filter by name..." class="form-control-static filter-input" v-model="filterString" />
+    </div>
+  </div>
+</template>
+
+<style scoped>
+.filter-group.filter-monitoring {
+  width: 100%;
+  margin-top: 0.5em;
+}
+
+.filter-group {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  width: 50%;
+  position: relative;
+  top: -3px;
+  margin-top: -26px;
+}
+
+.filter-group > *:not(:first-child) {
+  margin-left: 1.5em;
+}
+
+.filter-input input {
+  display: inline-block;
+  width: 100%;
+  padding-right: 10px;
+  padding-left: 30px;
+  border: 1px solid #aaa;
+  border-radius: 4px;
+}
+
+div.filter-input {
+  position: relative;
+  width: 280px;
+}
+
+.filter-input:before {
+  font-family: "FontAwesome";
+  width: 1.43em;
+  content: "\f0b0";
+  color: #919e9e;
+  position: absolute;
+  top: calc(50% - 0.7em);
+  left: 0.75em;
+}
+</style>
