+++ src/ServicePulse.Host/vue/src/components/monitoring/MonitoringHistoryPeriod.vue
@@ -27,11 +27,13 @@ async function selectHistoryPeriod(period) {
   </ul>
 </template>
 
-<style>
-.nav-pills.period-selector {
-  display: inline-flex;
-  position: relative;
-  top: 30px;
+<style scoped>
+.period-selector {
+  color: #00a3c4;
+}
+
+.nav li {
+  display: flex;
 }
 
 .nav-pills.period-selector > li.active > a,
@@ -41,7 +43,7 @@ async function selectHistoryPeriod(period) {
   font-weight: bold;
   background-color: initial;
   border-bottom: 3px solid #000;
-  padding-bottom: 10px;
+  padding: 10px 6px;
 }
 
 .nav-pills.period-selector > li > a:hover {
